backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Ricardo Sprocati",email = "ricardosprocatioliveira@outlook.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django (>=5.0,<6.0)",
    "djangorestframework (>=3.15,<4.0)",
    "djangorestframework-simplejwt (>=5.3,<6.0)",
    "drf-spectacular (>=0.27,<0.28)",
    "pillow (>=10.4,<11.0)",
    "django-cors-headers (>=4.4,<5.0)",
    "python-dotenv (>=1.0,<2.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-django = "^4.11.1"
model-bakery = "^1.20.5"
black = "^25.1.0"
isort = "^6.0.1"
ruff = "^0.12.8"
coverage = "^7.10.2"

